refactor: rename validator module to github namespace
