+++ core/src/main/java/io/lonmstalker/tgkit/core/wizard/StepBuilderImpl.java
@@ -18,10 +18,10 @@ package io.lonmstalker.tgkit.core.wizard;
 import io.lonmstalker.tgkit.core.BotRequest;
 import io.lonmstalker.tgkit.core.i18n.MessageKey;
 import io.lonmstalker.tgkit.core.validator.Validator;
-import io.lonmstalker.tgkit.validator.impl.LocationValidators;
-import io.lonmstalker.tgkit.validator.impl.PhotoValidators;
-import io.lonmstalker.tgkit.validator.impl.TextValidators;
-import io.lonmstalker.tgkit.validator.impl.VideoValidators;
+import io.github.tgkit.validator.impl.LocationValidators;
+import io.github.tgkit.validator.impl.PhotoValidators;
+import io.github.tgkit.validator.impl.TextValidators;
+import io.github.tgkit.validator.impl.VideoValidators;
 import java.time.Duration;
 import java.util.List;
 import java.util.function.BiConsumer;
